Lab 5

+++ bprint.py
@@ -1,10 +1,19 @@
 from flask import Blueprint, jsonify, request
 import dbu
 from models import Session, user_table, car_table, order_table
-from schema import UserDetails, UserQuery, OrderDetails, OrderQuery, CarDetails, CarQuery, LoginData, Token, ListUsersReq, Response
+from schema import UserDetails, UserQuery, OrderDetails, OrderQuery, CarDetails, CarQuery, LoginData, \
+    ListUsersReq, Response
 from contextlib import contextmanager
+from flask_jwt_extended import jwt_required, create_access_token, get_jwt_identity
+import datetime
+import bcrypt
 
 api_blueprint = Blueprint('api', __name__)
+
+
+def set_password(self, pw):
+    pwhash = bcrypt.hashpw(pw.encode('utf8'), bcrypt.gensalt())
+    self.password_hash = pwhash.decode('utf8')
 
 
 @contextmanager
@@ -25,35 +34,65 @@ def session_scope():
 
 
 @api_blueprint.route("/login", methods=["POST"])
+@jwt_required
 def login():
-    LoginData().load(request.json)
-    return jsonify(Token().dump({"token": ""}))
+    from app import bcrypt
+    data = LoginData().load(request.json)
+    if data:
+        user = dbu.get_entry_by_username(user_table, username=data["username"])
+        hpw = bcrypt.generate_password_hash(data["password"])
+        if not user:
+            return jsonify({"message": "Couldn't find user!"})
+        if bcrypt.check_password_hash(hpw, data["password"]):
+            access_token = create_access_token(identity=data["username"], expires_delta=datetime.timedelta(days=365))
+            return jsonify(access_token=access_token, id=user.id), 200
+        elif not data["password"] == user.password:
+            return jsonify({"message": "Wrong pass!", "data pass": data["password"], "user pass": user.password})
+    elif not data:
+        return jsonify({"message": "Wrong data!"})
 
 
 @api_blueprint.route("/user", methods=["GET"])
+@jwt_required
 def list_users():
     with session_scope():
-        args = ListUsersReq().load(request.args)
-        userlist = dbu.list_users(args.get("email"), args.get("username"))
-        return jsonify(UserDetails(many=True).dump(userlist))
+        current_user = get_jwt_identity()
+        user = dbu.get_entry_by_username(user_table, current_user)
+        if user.admin:
+            args = ListUsersReq().load(request.args)
+            userlist = dbu.list_users(args.get("email"), args.get("username"))
+            return jsonify(UserDetails(many=True).dump(userlist))
+        else:
+            return jsonify({"code": "401", "type": "UNAUTHORISED_ACCESS"})
 
 
 @api_blueprint.route("/user", methods=["POST"])
+@jwt_required
 def create_user():
     with session_scope():
+        from app import bcrypt
         user_details = UserQuery().load(request.get_json(force=True))
+        user_details["password"] = bcrypt.generate_password_hash(user_details["password"]).decode('UTF-8')
         user = dbu.create_entry(user_table, **user_details)
-        return jsonify({"id": UserDetails().dump(user)["id"]})
+        access_token = create_access_token(identity=user.username, expires_delta=datetime.timedelta(days=365))
+        return jsonify(access_token=access_token, id=UserDetails().dump(user)["id"]), 200
 
 
 @api_blueprint.route("/user/<int:id>", methods=["GET"])
+@jwt_required
 def user_by_id(id):
     with session_scope():
-        user = dbu.get_entry_by_id(user_table, id)
-        return jsonify(UserDetails().dump(user))
+        current_user = get_jwt_identity()
+        user = dbu.get_entry_by_username(user_table, current_user)
+        if user.admin:
+            user = dbu.get_entry_by_id(user_table, id)
+            return jsonify(UserDetails().dump(user))
+        else:
+            return jsonify({"code": "401", "type": "UNAUTHORISED_ACCESS"})
 
 
 @api_blueprint.route("/user/<int:id>", methods=["PUT"])
+@jwt_required
 def update_user(id):
     with session_scope():
         user_details = UserQuery().load(request.json)
@@ -63,6 +102,7 @@ def update_user(id):
 
 
 @api_blueprint.route("/user/<int:id>", methods=["DELETE"])
+@jwt_required
 def delete_user(id):
     with session_scope():
         dbu.delete_entry(user_table, id)
@@ -70,6 +110,7 @@ def delete_user(id):
 
 
 @api_blueprint.route("/cars", methods=["GET"])
+@jwt_required
 def get_inventory():
     with session_scope():
         cars = dbu.list_cars()
@@ -77,6 +118,7 @@ def get_inventory():
 
 
 @api_blueprint.route("/cars/car/<int:carId>", methods=["GET"])
+@jwt_required
 def get_car_by_id(carId):
     with session_scope():
         car = dbu.get_car_by_id(car_table, carId)
@@ -84,30 +126,49 @@ def get_car_by_id(carId):
 
 
 @api_blueprint.route("/cars/car", methods=["POST"])
+@jwt_required
 def create_car():
     with session_scope():
-        car_details = CarQuery().load(request.json)
-        car = dbu.create_entry(car_table, **car_details)
-        return jsonify({"carId": CarDetails().dump(car)["carId"]})
+        current_user = get_jwt_identity()
+        user = dbu.get_entry_by_username(user_table, current_user)
+        if user.admin:
+            car_details = CarQuery().load(request.json)
+            car = dbu.create_entry(car_table, **car_details)
+            return jsonify({"carId": CarDetails().dump(car)["carId"]})
+        else:
+            return jsonify(code=401, type="UNAUTHORISED_ACCESS")
 
 
 @api_blueprint.route("/cars/car/<int:carId>", methods=["PUT"])
+@jwt_required
 def update_car(carId):
     with session_scope():
-        car_details = CarQuery().load(request.json)
-        car = dbu.get_car_by_id(car_table, carId)
-        dbu.update_entry(car, **car_details)
-        return jsonify(Response().dump({"code": "200"}))
+        current_user = get_jwt_identity()
+        user = dbu.get_entry_by_username(user_table, current_user)
+        if user.admin:
+            car_details = CarQuery().load(request.json)
+            car = dbu.get_car_by_id(car_table, carId)
+            dbu.update_entry(car, **car_details)
+            return jsonify(Response().dump({"code": "200"}))
+        else:
+            return jsonify(code=401, type="UNAUTHORISED_ACCESS")
 
 
 @api_blueprint.route("/cars/car/<int:carId>", methods=["DELETE"])
+@jwt_required
 def delete_car(carId):
     with session_scope():
-        dbu.delete_car(car_table, carId)
-        return jsonify(Response().dump({"code": "200"}))
+        current_user = get_jwt_identity()
+        user = dbu.get_entry_by_username(user_table, current_user)
+        if user.admin:
+            dbu.delete_car(car_table, carId)
+            return jsonify(Response().dump({"code": "200"}))
+        else:
+            return jsonify(code=401, type="UNAUTHORISED_ACCESS")
 
 
 @api_blueprint.route("/cars/car/<int:carId>/order", methods=["POST"])
+@jwt_required
 def place_order(carId):
     with session_scope():
         order_data = OrderQuery().load(request.json)
@@ -122,6 +183,7 @@ def place_order(carId):
 
 
 @api_blueprint.route("/cars/car/<int:carId>/order/<int:orderId>", methods=["GET"])
+@jwt_required
 def get_order_by_id(carId, orderId):
     with session_scope():
         order = dbu.get_entry_by_id(order_table, id=orderId)
@@ -129,6 +191,7 @@ def get_order_by_id(carId, orderId):
 
 
 @api_blueprint.route("/cars/car/<int:carId>/order/<int:orderId>", methods=["DELETE"])
+@jwt_required
 def delete_order(carId, orderId):
     with session_scope():
         dbu.delete_entry(order_table, id=orderId)
